+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * GLM binary classification in the doctor
 * Elastic Net Penalization
 * Sample Weighting
-* Offset and Exposure Support
+* Multi Offset and Exposure Support
 * Spline Regression as a recipe
 * Spline Regression as a prepare step
 * GLM Analysis View including metrics and Actual vs Predicted graphs in Analysis and Saved Models